Extract compiling to individual trait for re-usability
